miden_assembly::ast

Module constants

Modules

eval

Structs

Constant

Represents a constant definition in Miden Assembly syntax, i.e. const.FOO = 1 + 1.

Enums

ConstEvalError

An error raised during evaluation of a constant expression

ConstantExpr

Represents a constant expression or value in Miden Assembly syntax.

ConstantOp

Represents the set of binary arithmetic operators supported in Miden Assembly syntax.

ConstantValue

Represents a constant value in Miden Assembly syntax.

HashKind

Represents the type of the final value to which some string value should be converted.

Traits

ConstEnvironment

There are two phases to constant evaluation, one during semantic analysis, and another phase performed during linking of the final assembly, on any constant expressions that were left partially or unevaluated during semantic analysis due to external references. This trait is used to abstract over the environment in which the evaluator runs, so that we can use it in both phases by simply providing a suitable implementation.