miden_assembly::ast

Module path

Modules

arbitrary

Structs

Iter

Returns an iterator over the path components represented in the provided source.

Path

A borrowed reference to a subset of a path, e.g. another Path or a PathBuf

PathBuf

Path to an item in a library, i.e. module, procedure, constant or type.

Enums

NamespaceError

Represents an error when parsing or validating a library namespace

PathComponent

Represents a single component of a Path

PathError

Represents errors that can occur when creating, parsing, or manipulating Paths

Traits

Join

This trait is used to implement joining of a path or path component to a Path or PathBuf.

StartsWith

This trait abstracts over the concept of matching a prefix pattern against a path

Functions

deserialize

Deserialize a Path-like value

deserialize_spanned

Deserialize a Path-like value wrapped in a Span

serialize

Serialize a Path-like value