miden_assembly::ast::path

Module arbitrary

Functions

bare_path_random_length

Generate a Arc<Path> of random length, but at least min components.

bare_pathbuf

A strategy to produce a PathBuf of up between min and max bare components.

bare_pathbuf_random_length

Generate a PathBuf of random length, but at least min components.

builtin_type_pathbuf

A strategy to produce a PathBuf corresponding to a built-in type reference.

constant_path_random_length

Generate a Arc<Path> of random length, but at least min prefix components, that is valid for use with constant items.

constant_pathbuf

A strategy to produce a PathBuf of between min and max prefix components (either bare or quoted), with a constant identifier as the last component.

constant_pathbuf_random_length

Generate a PathBuf of random length, but at least min prefix components, that is valid for use with constant items.

path_random_length

Generate a Arc<Path> of random length, but at least min components.

pathbuf

A strategy to produce a PathBuf of between min and max components (either bare or quoted).

pathbuf_random_length

Generate a PathBuf of random length, but at least min components.

type_path_random_length

Generate a Arc<Path> of random length, but at least min prefix components, that is valid for use with type references.

type_pathbuf

A strategy to produce a PathBuf corresponding to a valid TypeExpr::Ref, where user-defined type paths will have between min and max components.

type_pathbuf_random_length

Generate a PathBuf of random length, but at least min prefix components, that is valid for use as a type reference.

user_defined_type_path_random_length

Generate a Arc<Path> of random length, but at least min prefix components, that is valid for use with user-defined type items.

user_defined_type_pathbuf

A strategy to produce a PathBuf of up to between min and max prefix components (either bare or quoted), with a user-defined type identifier as the last component.

user_defined_type_pathbuf_random_length

Generate a PathBuf of random length, but at least min prefix components, that is valid for use with user-defined type items.