miden_assembly::ast

Type Alias ImmFelt

pub type ImmFelt = Immediate<Felt>;

A field element immediate

Aliased Type

pub enum ImmFelt {
    Value(Span<Felt>),
    Constant(Ident),
}

Variants

Value(Span<Felt>)

A literal integer value, either decimal or hex-encoded

Constant(Ident)

A constant identifier

This must refer to a constant definition in the current module.

All immediates of this type are folded to Value during semantic analysis, once all constant definitions are evaluated.