miden_assembly::ast

Crate types

Structs

ArrayType

A fixed-size, homogenous vector type.

EnumType

An enum type is a special type that takes one of two forms:

FunctionType

This represents the type of a function, i.e. it’s parameters and results, and expected calling convention.

NameAndType

A wrapper around Type which permits a Type to be optionally named.

PointerType

A pointer to an object in memory

StructField

This represents metadata about a field of a StructType

StructType

This represents a structured aggregate type

Variant

Enums

AddressSpace

The address space a pointer address is evaluated in.

CallConv

Represents the calling convention of a function.

InvalidAddressSpaceError

This error is raised when parsing an AddressSpace

InvalidEnumTypeError

The error type returned when attempting to construct an invalid EnumType

Type

Represents the type of a value in the HIR type system

TypeRepr

This represents metadata about how a structured type will be represented in memory

Traits

Alignable

This trait represents an alignable primitive integer value representing an address