miden_assembly::ast::types

Enum Type

pub enum Type {

    Unknown,
    Never,
    I1,
    I8,
    U8,
    I16,
    U16,
    I32,
    U32,
    I64,
    U64,
    I128,
    U128,
    U256,
    F64,
    Felt,
    Ptr(Arc<PointerType>),
    Struct(Arc<StructType>),
    Enum(Arc<EnumType>),
    Array(Arc<ArrayType>),
    List(Arc<Type>),
    Function(Arc<FunctionType>),

}

Represents the type of a value in the HIR type system

Variants

Unknown

This indicates a failure to type a value, or a value which is untypable

Never

This type is the bottom type, and represents divergence, akin to Rust’s Never/! type

I1

A 1-bit integer, i.e. a boolean value.

When the bit is 1, the value is true; 0 is false.

I8

An 8-bit signed integer.

U8

An 8-bit unsigned integer.

I16

A 16-bit signed integer.

U16

A 16-bit unsigned integer.

I32

A 32-bit signed integer.

U32

A 32-bit unsigned integer.

I64

A 64-bit signed integer.

U64

A 64-bit unsigned integer.

I128

A 128-bit signed integer.

U128

A 128-bit unsigned integer.

U256

A 256-bit unsigned integer.

F64

A 64-bit IEEE-754 floating-point value.

NOTE: These are currently unsupported in practice, but is reserved here for future use.

Felt

A field element corresponding to the native Miden field (currently the Goldilocks field)

Ptr(Arc<PointerType>)

A pointer to a value in a byte-addressable address space.

Pointers of this type are not equivalent to element addresses as referred to in the Miden Assembly documentation, but do have a straightforward conversion.

Struct(Arc<StructType>)

A compound type of fixed shape and size

Enum(Arc<EnumType>)

A tagged type enumeration with a fixed number of variants

Array(Arc<ArrayType>)

A vector of fixed size

List(Arc<Type>)

A dynamically sized list of values of the given type

NOTE: Currently this only exists to support the Wasm Canonical ABI, but it has no defined represenation yet, so in practice cannot be used in most places except during initial translation in the Wasm frontend.

Function(Arc<FunctionType>)

A reference to a function with the given type signature

Implementations

impl Type

pub fn to_raw_parts(self) -> Option<SmallVec<[Type; 4]>>

Convert this type into a vector of types corresponding to how this type will be represented in memory.

The largest “part” size is 32 bits, so types that fit in 32 bits remain unchanged. For types larger than 32 bits, they will be broken up into parts that do fit in 32 bits, preserving accurate types to the extent possible. For types smaller than 32 bits, they will be merged into packed structs no larger than 32 bits, to preserve the type information, and make it possible to reason about how to extract parts of the original type.

For an example, a struct of type { *ptr, u8, u8 } will be encoded on the operand stack as [*ptr, {u8, u8}], where the first value is the 32-bit pointer field, and the remaining fields are encoded as a 16-bit struct in the second value.

pub fn split(self, n: usize) -> (Type, Option<Type>)

Split this type into two parts:

The first part is no more than n bytes in size, and may contain the type itself if it fits

The second part is None if the first part is smaller than or equal in size to the requested split size

The second part is Some if there is data left in the original type after the split. This part will be a type that attempts to preserve, to the extent possible, the original type structure, but will fall back to an array of bytes if a larger type must be split down the middle somewhere.

pub fn min_alignment(&self) -> usize

Returns the minimum alignment, in bytes, of this type

pub fn size_in_bits(&self) -> usize

Returns the size in bits of this type, without alignment padding.

pub fn size_in_bytes(&self) -> usize

Returns the minimum number of bytes required to store a value of this type

pub fn aligned_size_in_bytes(&self) -> usize

Same as size_in_bytes, but with sufficient padding to guarantee alignment of the value.

pub fn size_in_felts(&self) -> usize

Returns the size in field elements of this type

pub fn size_in_words(&self) -> usize

Returns the size in words of this type

pub fn layout(&self) -> Layout

Returns the layout of this type in memory

pub fn is_loadable(&self) -> bool

Returns true if this type can be loaded on to the operand stack

The rule for “loadability” is a bit arbitrary, but the purpose is to force users of the IR to either pass large values by reference, or calculate the addresses of the individual fields needed from a large structure or array, and issue loads/stores against those instead.

In effect, we reject loads of values that are larger than a single word, as that is the largest value which can be worked with on the operand stack of the Miden VM.

impl Type

pub fn is_zst(&self) -> bool

Returns true if this type is a zero-sized type, which includes:

Types with no size, e.g. Never

Zero-sized arrays

Arrays with a zero-sized element type

Structs composed of nothing but zero-sized fields

pub fn is_numeric(&self) -> bool

Returns true if this type is any numeric type

pub fn is_integer(&self) -> bool

Returns true if this type is any integral type

pub fn is_signed_integer(&self) -> bool

Returns true if this type is any signed integral type

pub fn is_unsigned_integer(&self) -> bool

Returns true if this type is any unsigned integral type

pub fn as_unsigned(&self) -> Type

Get this type as its unsigned integral twin, e.g. i32 becomes u32.

This function will panic if the type is not an integer type, or has no unsigned representation

pub fn as_signed(&self) -> Type

Get this type as its signed integral twin, e.g. u32 becomes i32.

This function will panic if the type is not an integer type, or has no signed representation

pub fn is_float(&self) -> bool

Returns true if this type is a floating-point type

pub fn is_felt(&self) -> bool

Returns true if this type is the field element type

pub fn is_pointer(&self) -> bool

Returns true if this type is a pointer type

pub fn pointee(&self) -> Option<&Type>

Returns the type of the pointee, if this type is a pointer type

pub fn is_struct(&self) -> bool

Returns true if this type is a struct type

pub fn is_array(&self) -> bool

Returns true if this type is an array type

pub fn is_list(&self) -> bool

Returns true if this type is a dynamically-sized vector/list type

pub fn is_function(&self) -> bool

Returns true if this type is a function reference type