miden_assembly

Module library

Structs

ConstantExport

ConstantInfo

Stores the name and value of a constant

KernelLibrary

Represents a library containing a Miden VM kernel.

Library

Represents a library where all modules were compiled into a MastForest.

ModuleInfo

ProcedureExport

ProcedureInfo

Stores the name and digest of a procedure.

TypeExport

TypeInfo

Stores the name and concrete type of a type declaration

Version

SemVer version as defined by https://semver.org.

VersionError

Error parsing a SemVer version or version requirement.

Enums

ItemInfo

Stores information about an item

LibraryError

LibraryExport

Metadata about a procedure exported by the interface of a Library