miden_assembly::mast

Function build_node_with_remapped_ids

pub fn build_node_with_remapped_ids<NMap, DMap>(
    node_id: MastNodeId,
    node: MastNode,
    source_forest: &MastForest,
    node_remapping: &NMap,
    decorator_remapping: &DMap,
) -> Result<MastNodeBuilder, MastForestError>
where
    NMap: LookupByIdx<MastNodeId, MastNodeId>,
    DMap: LookupByIdx<DecoratorId, DecoratorId>,

Builds a node builder with remapped children and decorators using the provided mappings.

This is a generic helper used by both MastForestMerger and MastForestBuilder to avoid code duplication when copying nodes between forests.