miden_assembly

Struct PathBuf

pub struct PathBuf { /* private fields */ }

Path to an item in a library, i.e. module, procedure, constant or type.

Implementations

impl PathBuf

Constructors

pub fn with_capacity(capacity: usize) -> PathBuf

Get an empty PathBuf with capacity bytes allocated for the underlying path storage

pub fn new<S>(source: &S) -> Result<PathBuf, PathError>
where S: AsRef<str> + ?Sized,

Returns a new path created from the provided source.

A path consists of at list of components separated by :: delimiter. A path must contain at least one component.

Errors

Returns an error if:

The path is empty.

Any component of the path is empty.

Any component is not a valid identifier (quoted or unquoted) in Miden Assembly syntax, i.e. starts with an ASCII alphabetic character, contains only printable ASCII characters, except for ::, which must only be used as a path separator.

pub fn absolute<S>(source: &S) -> PathBuf
where S: AsRef<str> + ?Sized,

Create an absolute Path from a pre-validated string

pub fn relative<S>(source: &S) -> PathBuf
where S: AsRef<str> + ?Sized,

Create a relative Path from a pre-validated string

pub fn as_path(&self) -> &Path

Get a Path corresponding to the this PathBuf

pub fn into_boxed_path(self) -> Box<Path>

Convert this mutable PathBuf into an owned, read-only alloc::boxed::Box<Path>

impl PathBuf

Mutation

pub fn set_parent<P>(&mut self, parent: &P)
where P: AsRef<Path> + ?Sized,

Overrides the parent prefix of this path.

The parent prefix is the part of the path consisting of all components but the last one.

If there is only a single component in self, this function is equivalent to appending self to parent.

pub fn push<P>(&mut self, path: &P)
where P: AsRef<Path> + ?Sized,

Extends self with path

If path is absolute, it replaces the current path.

This function ensures that the joined path correctly delimits each path component, and that each component is in canonical form.

pub fn push_component<S>(&mut self, component: &S)
where S: AsRef<str> + ?Sized,

Extends self with component.

Unlike Self::push, which appends another Path to the buffer - this method appends the given string as a single path component, ensuring that the content is quoted properly if needed.

If :: is pushed, it is treated as a literal root component (i.e. it makes the path absolute). On a non-empty PathBuf, this has the effect of clearing the path, similar to what happens if you call PathBuf::push with an absolute path.

Pushing components using this method guarantees they are in canonical form.

pub fn pop(&mut self) -> bool

Truncates self to Path::parent.

Returns false if self.parent() is None, otherwise true.