miden_assembly

Struct Span

pub struct Span<T> { /* private fields */ }

This type is used to wrap any T with a SourceSpan, and is typically used when it is not convenient to add a SourceSpan to the type - most commonly because we don’t control the type.

Implementations

impl<T> Span<T>

pub fn from_serde_spanned(source_id: SourceId, spanned: Spanned<T>) -> Span<T>

impl<T> Span<T>

pub fn new(span: impl Into<SourceSpan>, spanned: T) -> Span<T>

Creates a span for spanned with span.

pub fn at(source_id: SourceId, offset: usize, spanned: T) -> Span<T>

Creates a span for spanned representing a single location, offset.

pub fn unknown(spanned: T) -> Span<T>

Creates a Span from a value with an unknown/default location.

pub fn with_span(self, span: SourceSpan) -> Span<T>

Consume this Span and get a new one with span as the underlying source span

pub const fn span(&self) -> SourceSpan

Gets the associated SourceSpan for this spanned item.

pub const fn inner(&self) -> &T

Gets a reference to the spanned item.

pub fn map<U, F>(self, f: F) -> Span<U>
where F: FnMut(T) -> U,

Applies a transformation to the spanned value while retaining the same SourceSpan.

pub fn as_deref<U>(&self) -> Span<&U>
where T: Deref<Target = U>, U: ?Sized,

Like Option<T>::as_deref, this constructs a Span<U> wrapping the result of dereferencing the inner value of type T as a value of type U.

pub fn as_ref(&self) -> Span<&T>

Gets a new Span that borrows the inner value.

pub fn set_source_id(&mut self, id: SourceId)

Manually set the source id for the span of this item

See also SourceSpan::set_source_id.

pub fn shift(&mut self, count: ByteOffset)

Shifts the span right by count units

pub fn extend(&mut self, count: ByteOffset)

Extends the end of the span by count units.

pub fn into_parts(self) -> (SourceSpan, T)

Consumes this span, returning the component parts, i.e. the SourceSpan and value of type T.

pub fn into_inner(self) -> T

Unwraps the spanned value of type T.

impl<T> Span<T>
where T: Serializable,

pub fn write_into_with_options<W>(&self, target: &mut W, debug: bool)
where W: ByteWriter,

impl<T> Span<T>
where T: Deserializable,

pub fn read_from_with_options<R>( source: &mut R, debug: bool, ) -> Result<Span<T>, DeserializationError>
where R: ByteReader,